Minor change to fix AUTOCOMMIT logic
By default the driver sets AUTOCOMMIT to ON which could lead to errors connecting to a non-logged database such as "Driver not capable. SQLCODE=-11092".

The behaviour can be controlled setting the connection attribute (SQL_ATTR_AUTOCOMMIT), e.g.

   conn =  IfxPy.connect(ConStr, "", "",{ IfxPy.SQL_ATTR_AUTOCOMMIT : IfxPy.SQL_AUTOCOMMIT_OFF })
 
The change here is to  silently ignore the error the driver may get when trying to roll-back any open transactions if it was connected to a non logged database.
In theory the roll-back shouldn't be there at all, as is only executed if AUTOCOMMIT is OFF, and if is OFF, it means the application should do all commits/rollbacks, but leaving it there won't do much harm.